+++ src/main/java/com/example/springboot/services/BookingService.java
@@ -12,11 +12,11 @@ import java.util.List;
 
 public interface BookingService {
     List<BookingDTO> getAll();
-    BookingDTO getById(int id) throws ItemNotFoundException;
+    BookingDTO getById(long id) throws ItemNotFoundException;
     List<BookingDTO> getByUser(User user);
     List<Car> selCarsByDateBooking(LocalDate start, LocalDate end);
     void insBooking(BookingDTO request) throws ItemNotFoundException;
     void editBooking(BookingDTO request) throws ItemNotFoundException;
     void newBooking(BookingDTO booking) throws BadRequestException, ItemNotFoundException;
-    void delBooking(int id) throws ItemNotFoundException;
+    void delBooking(long id) throws ItemNotFoundException;
 }
